refactor: extract QuarkusControllerConfigurationBuilder to simplify
